requirements.txt
web3==5.29.0
pytest==7.1.2
pytest-random-order==1.0.4
assertpy==1.1
pytest-html==3.1.1
allure-pytest==2.9.45